desktop_bridge/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.32.1
pywin32>=311,<312
pystray==0.19.5
Pillow>=10.0.0,<12
slowapi==0.1.9
pytest>=8.0
httpx>=0.27
